Add New Comment on your post notifications
This notification is sent only to the post owner, unless this person turned off the notifications.
A simple first step into the notifications by mail world :)

The mail contains a link to user profile
The link to the user profile is required to give an easy access to
notifications disabling.

Also having a preview for the notify_post_owner method

We can `Comment.first` because it is part of the db:seeds method. So
there should, in development, always be one.